Fix window lifecycle project wiring
